# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
  model.Account:
    properties:
      createdAt:
        description: Creation timestamp
        type: integer
      down:
        description: Total downloaded traffic in bytes
        type: integer
      enable:
        description: Whether the account is enabled
        type: boolean
      expiryTime:
        description: Expiration timestamp (0 = never expires)
        type: integer
      id:
        type: integer
      remark:
        description: Account description
        type: string
      reset:
        description: Traffic reset period in days (0 = never)
        type: integer
      subId:
        description: Subscription UUID
        type: string
      tgId:
        description: Telegram user ID for notifications
        type: integer
      totalGB:
        description: Total traffic limit in GB (0 = unlimited)
        type: integer
      up:
        description: Total uploaded traffic in bytes
        type: integer
      updatedAt:
        description: Last update timestamp
        type: integer
      username:
        description: Unique account identifier
        type: string
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: API documentation for 3X-UI Panel (Master-Slave Cluster)
  title: 3X-UI Cluster API
  version: "1.0"
paths:
  /panel/api/account/{id}/clients:
    get:
      description: Returns all clients linked to an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get account clients
      tags:
      - Accounts
  /panel/api/account/{id}/clients/add:
    post:
      consumes:
      - application/json
      description: Associates an inbound client with an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Add client to account
      tags:
      - Accounts
  /panel/api/account/{id}/clients/remove/{clientEmail}:
    post:
      description: Disassociates a client from an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Client email
        in: path
        name: clientEmail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Remove client from account
      tags:
      - Accounts
  /panel/api/account/{id}/traffic:
    get:
      description: Returns aggregated upload/download traffic for an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get account traffic
      tags:
      - Accounts
  /panel/api/account/add:
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Add account
      tags:
      - Accounts
  /panel/api/account/del/{id}:
    post:
      description: Deletes an account and all associated client bindings
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Delete account
      tags:
      - Accounts
  /panel/api/account/get/{id}:
    get:
      description: Returns a single account by its ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get account
      tags:
      - Accounts
...
  /panel/api/account/reset/traffic/{id}:
    post:
      description: Resets all traffic counters for an account and pushes config to
        slaves
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Reset account traffic
      tags:
      - Accounts
  /panel/api/account/update/{id}:
    post:
      consumes:
      - application/json
      description: Updates an existing account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Update account
      tags:
      - Accounts
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: 3x-ui
    type: apiKey
swagger: "2.0"